Move `NativeExecutorContext`'s dependencies under `context` module
