+++ crates/contracts/ab-contracts-executor/src/context/aligned_buffer.rs
