+++ crates/contracts/ab-contracts-executor/src/context/aligned_buffer/tests.rs
@@ -1,4 +1,4 @@
-use crate::aligned_buffer::{AlignedBytes, OwnedAlignedBuffer};
+use crate::context::aligned_buffer::{AlignedBytes, OwnedAlignedBuffer};
 
 const EXPECTED_ALIGNMENT: usize = size_of::<AlignedBytes>();
